# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GatherUp API - Community events platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /push/subscribe:
    post:
      operationId: subscribePush
      tags: [push]
      summary: Register a Web Push subscription for the logged-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionInput"
      responses:
        "204":
          description: Subscribed
        "401":
          description: Not authenticated
  /push/unsubscribe:
    post:
      operationId: unsubscribePush
      tags: [push]
      summary: Remove a Web Push subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnsubscribeInput"
      responses:
        "204":
          description: Unsubscribed
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    PushSubscriptionInput:
      type: object
      required: [endpoint, keys]
      properties:
        endpoint:
          type: string
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh:
              type: string
            auth:
              type: string
    PushUnsubscribeInput:
      type: object
      required: [endpoint]
      properties:
        endpoint:
          type: string