# lib/api-zod/src/generated/types/pushSubscriptionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GatherUp API - Community events platform
 * OpenAPI spec version: 0.1.0
 */
import type { PushSubscriptionInputKeys } from './pushSubscriptionInputKeys';

export interface PushSubscriptionInput {
  endpoint: string;
  keys: PushSubscriptionInputKeys;
}