# lib/api-zod/src/generated/types/pushUnsubscribeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GatherUp API - Community events platform
 * OpenAPI spec version: 0.1.0
 */

export interface PushUnsubscribeInput {
  endpoint: string;
}